Knn fashion mnist (#14)
* Fashion mnist

* report

* refactor MLP and KNN

* report conclusion
