+++ docs/fripes.md
@@ -1,0 +1,76 @@
+# Résultats
+## KNN
+Nb imagettes : 1000
+Pourcentage de reussite : 4%
+Nombre de reussite : 42
+Nombre d'echec : 958
+
+## MLP
+### Configuration 1
+- Nb imagettes : 1000
+- Fonction d'activation : tanh
+- Neurones :
+  - Entrées : 784
+  - Cachées : [9]
+  - Sorties : 10
+- Accuracy: 57.3%
+### Configuration 2
+- Nb imagettes : 1000
+- Fonction d'activation : sigmoid
+- Neurones :
+  - Entrées : 784
+  - Cachées : [9]
+  - Sorties : 10
+- Accuracy: 59.0%
+### Configuration 3
+- Nb imagettes : 1000
+- Fonction d'activation : tanh
+- Neurones :
+  - Entrées : 784
+  - Cachées : [100]
+  - Sorties : 10
+- Accuracy: 24.7%
+### Configuration 4
+- Nb imagettes : 1000
+- Fonction d'activation : sigmoid
+- Neurones :
+  - Entrées : 784
+  - Cachées : [100]
+  - Sorties : 10
+- Accuracy: 38.9%
+### Configuration 5
+- Nb imagettes : 1000
+- Fonction d'activation : tanh
+- Neurones :
+  - Entrées : 784
+  - Cachées : [250]
+  - Sorties : 10
+- Accuracy: 23.9%
+### Configuration 6
+- Nb imagettes : 1000
+- Fonction d'activation : sigmoid
+- Neurones :
+  - Entrées : 784
+  - Cachées : [250]
+  - Sorties : 10
+- Accuracy: 32.300000000000004%
+### Configuration 7
+- Nb imagettes : 1000
+- Fonction d'activation : tanh
+- Neurones :
+  - Entrées : 784
+  - Cachées : [100, 100]
+  - Sorties : 10
+- Accuracy: 28.299999999999997%
+### Configuration 8
+- Nb imagettes : 1000
+- Fonction d'activation : sigmoid
+- Neurones :
+  - Entrées : 784
+  - Cachées : [100, 100]
+  - Sorties : 10
+- Accuracy: 9.5%
+
+## Conclusion
+Cette fois, le taux de réussite est extrêmement faible pour KNN, et cela en raison de la ressemblance entre tous les vêtements.
+Pour le MLP, les résultats restent similaires par rapport à la base de données des chiffres manuscrits.
